chore: 키워드 관련 패키지 분리
