+++ src/main/java/in/koreatech/koin/domain/community/articles/exception/ArticleNotFoundException.java
@@ -1,0 +1,20 @@
+package in.koreatech.koin.domain.community.articles.exception;
+
+import in.koreatech.koin.global.exception.DataNotFoundException;
+
+public class ArticleNotFoundException extends DataNotFoundException {
+
+    private static final String DEFAULT_MESSAGE = "게시글이 존재하지 않습니다.";
+
+    public ArticleNotFoundException(String message) {
+        super(message);
+    }
+
+    public ArticleNotFoundException(String message, String detail) {
+        super(message, detail);
+    }
+
+    public static ArticleNotFoundException withDetail(String detail) {
+        return new ArticleNotFoundException(DEFAULT_MESSAGE, detail);
+    }
+}
