+++ src/main/java/in/koreatech/koin/domain/community/articles/model/Comment.java
@@ -1,0 +1,78 @@
+package in.koreatech.koin.domain.community.articles.model;
+
+import static jakarta.persistence.GenerationType.IDENTITY;
+import static lombok.AccessLevel.PROTECTED;
+
+import org.hibernate.annotations.Where;
+
+import in.koreatech.koin.global.domain.BaseEntity;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import jakarta.persistence.Transient;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Entity
+@Table(name = "comments")
+@Where(clause = "is_deleted=0")
+@NoArgsConstructor(access = PROTECTED)
+public class Comment extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
+    private Integer id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "article_id", nullable = false)
+    private Article article;
+
+    @NotNull
+    @Column(name = "content", nullable = false)
+    private String content;
+
+    @NotNull
+    @Column(name = "user_id", nullable = false)
+    private Integer userId;
+
+    @Size(max = 50)
+    @NotNull
+    @Column(name = "nickname", nullable = false, length = 50)
+    private String nickname;
+
+    @NotNull
+    @Column(name = "is_deleted", nullable = false)
+    private Boolean isDeleted = false;
+
+    @Transient
+    private boolean grantEdit = false;
+
+    @Transient
+    private boolean grantDelete = false;
+
+    public void updateAuthority(Integer userId) {
+        if (this.userId.equals(userId)) {
+            this.grantEdit = true;
+            this.grantDelete = true;
+        }
+    }
+
+    @Builder
+    private Comment(Article article, String content, Integer userId,
+        String nickname, Boolean isDeleted) {
+        this.article = article;
+        this.content = content;
+        this.userId = userId;
+        this.nickname = nickname;
+        this.isDeleted = isDeleted;
+    }
+}
