+++ src/main/java/in/koreatech/koin/domain/community/articles/repository/CommentRepository.java
@@ -1,0 +1,14 @@
+package in.koreatech.koin.domain.community.articles.repository;
+
+import java.util.List;
+
+import org.springframework.data.repository.Repository;
+
+import in.koreatech.koin.domain.community.articles.model.Comment;
+
+public interface CommentRepository extends Repository<Comment, Integer> {
+
+    List<Comment> findAllByArticleId(Integer articleId);
+
+    Comment save(Comment comment);
+}
