--- src/main/java/in/koreatech/koin/domain/community/controller/CommunityApi.java
@@ -1,119 +1,0 @@
-package in.koreatech.koin.domain.community.controller;
-
-import static in.koreatech.koin.domain.user.model.UserType.STUDENT;
-import static io.swagger.v3.oas.annotations.enums.ParameterIn.PATH;
-
-import java.util.List;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
-
-import in.koreatech.koin.domain.community.dto.ArticleKeywordCreateRequest;
-import in.koreatech.koin.domain.community.dto.ArticleKeywordResponse;
-import in.koreatech.koin.domain.community.dto.ArticleKeywordsResponse;
-import in.koreatech.koin.domain.community.dto.ArticleResponse;
-import in.koreatech.koin.domain.community.dto.ArticlesResponse;
-import in.koreatech.koin.domain.community.dto.HotArticleItemResponse;
-import in.koreatech.koin.global.auth.Auth;
-import in.koreatech.koin.global.auth.UserId;
-import in.koreatech.koin.global.ipaddress.IpAddress;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.validation.Valid;
-
-@Tag(name = "(Normal) Community: 커뮤니티", description = "커뮤니티 정보를 관리한다")
-public interface CommunityApi {
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(summary = "게시글 단건 조회")
-    @GetMapping("/articles/{id}")
-    ResponseEntity<ArticleResponse> getArticle(
-        @UserId Integer userId,
-        @Parameter(in = PATH) @PathVariable("id") Integer articleId,
-        @IpAddress String ipAddress
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(summary = "게시글 목록 조회")
-    @GetMapping("/articles")
-    ResponseEntity<ArticlesResponse> getArticles(
-        @RequestParam Integer boardId,
-        @RequestParam(required = false) Integer page,
-        @RequestParam(required = false) Integer limit
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(summary = "인기 게시글 목록 조회")
-    @GetMapping("/articles/hot")
-    ResponseEntity<List<HotArticleItemResponse>> getHotArticles();
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "400", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "409", content = @Content(schema = @Schema(hidden = true)))
-        }
-    )
-    @Operation(summary = "알림 키워드 생성")
-    @PostMapping("/articles/keyword")
-    ResponseEntity<ArticleKeywordResponse> createKeyword(
-        @Valid @RequestBody ArticleKeywordCreateRequest request,
-        @Auth(permit = {STUDENT}) Integer userId
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "400", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "409", content = @Content(schema = @Schema(hidden = true)))
-        }
-    )
-    @Operation(summary = "알림 키워드 삭제")
-    @DeleteMapping("/articles/keyword/{id}")
-    ResponseEntity<Void> deleteKeyword(
-        @PathVariable(value = "id") Integer keywordUserMapId,
-        @Auth(permit = {STUDENT}) Integer userId
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true)))
-        }
-    )
-    @Operation(summary = "자신의 알림 키워드 전체 조회")
-    @GetMapping("/articles/keyword/me")
-    ResponseEntity<ArticleKeywordsResponse> getMyKeywords(
-        @Auth(permit = {STUDENT}) Integer userId
-    );
-
-}
