+++ src/main/java/in/koreatech/koin/domain/community/keyword/repository/ArticleKeywordUserMapRepository.java
@@ -1,0 +1,29 @@
+package in.koreatech.koin.domain.community.keyword.repository;
+
+import java.util.List;
+import java.util.Optional;
+
+import org.springframework.data.repository.Repository;
+
+import in.koreatech.koin.domain.community.keyword.model.ArticleKeywordUserMap;
+import in.koreatech.koin.domain.community.keyword.exception.ArticleKeywordUserMapNotFoundException;
+
+public interface ArticleKeywordUserMapRepository extends Repository<ArticleKeywordUserMap, Integer> {
+
+    ArticleKeywordUserMap save(ArticleKeywordUserMap articleKeywordUserMap);
+
+    Long countByUserId(Integer userId);
+
+    void deleteById(Integer keywordId);
+
+    boolean existsByArticleKeywordId(Integer id);
+
+    Optional<ArticleKeywordUserMap> findById(Integer keywordUserMapId);
+
+    default ArticleKeywordUserMap getById(Integer keywordUserMapId) {
+        return findById(keywordUserMapId).orElseThrow(
+            () -> ArticleKeywordUserMapNotFoundException.withDetail("keywordUserMapId: " + keywordUserMapId));
+    }
+
+    List<ArticleKeywordUserMap> findAllByUserId(Integer userId);
+}
