+++ src/main/java/in/koreatech/koin/domain/community/keyword/service/KeywordService.java
@@ -1,0 +1,89 @@
+package in.koreatech.koin.domain.community.keyword.service;
+
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.Objects;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import in.koreatech.koin.domain.community.keyword.dto.ArticleKeywordCreateRequest;
+import in.koreatech.koin.domain.community.keyword.dto.ArticleKeywordResponse;
+import in.koreatech.koin.domain.community.keyword.dto.ArticleKeywordsResponse;
+import in.koreatech.koin.domain.community.keyword.exception.KeywordLimitExceededException;
+import in.koreatech.koin.domain.community.keyword.model.ArticleKeyword;
+import in.koreatech.koin.domain.community.keyword.model.ArticleKeywordUserMap;
+import in.koreatech.koin.domain.community.keyword.repository.ArticleKeywordRepository;
+import in.koreatech.koin.domain.community.keyword.repository.ArticleKeywordUserMapRepository;
+import in.koreatech.koin.domain.user.repository.UserRepository;
+import in.koreatech.koin.global.auth.exception.AuthorizationException;
+import in.koreatech.koin.global.exception.KoinIllegalArgumentException;
+import lombok.RequiredArgsConstructor;
+
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class KeywordService {
+
+    private final ArticleKeywordUserMapRepository articleKeywordUserMapRepository;
+    private final ArticleKeywordRepository articleKeywordRepository;
+    private final UserRepository userRepository;
+
+    @Transactional
+    public ArticleKeywordResponse createKeyword(Integer userId, ArticleKeywordCreateRequest request) {
+        String keyword = request.keyword().trim().toLowerCase();
+
+        if (keyword.contains(" ")) {
+            throw new KoinIllegalArgumentException("키워드에 공백을 포함할 수 없습니다.");
+        }
+
+        if (articleKeywordUserMapRepository.countByUserId(userId) >= 10) {
+            throw KeywordLimitExceededException.withDetail("userId: " + userId);
+        }
+
+        ArticleKeyword existingKeyword = articleKeywordRepository.findByKeyword(keyword)
+            .orElseGet(() -> articleKeywordRepository.save(
+                ArticleKeyword.builder()
+                    .keyword(keyword)
+                    .lastUsedAt(LocalDateTime.now())
+                    .build()
+            ));
+
+        ArticleKeywordUserMap keywordUserMap = ArticleKeywordUserMap.builder()
+            .user(userRepository.getById(userId))
+            .articleKeyword(existingKeyword)
+            .build();
+
+        existingKeyword.addUserMap(keywordUserMap);
+        articleKeywordUserMapRepository.save(keywordUserMap);
+
+        return new ArticleKeywordResponse(keywordUserMap.getId(), existingKeyword.getKeyword());
+    }
+
+    @Transactional
+    public void deleteKeyword(Integer userId, Integer keywordUserMapId) {
+        ArticleKeywordUserMap articleKeywordUserMap = articleKeywordUserMapRepository.getById(keywordUserMapId);
+        if (!Objects.equals(articleKeywordUserMap.getUser().getId(), userId)) {
+            throw AuthorizationException.withDetail("userId: " + userId);
+        }
+
+        articleKeywordUserMapRepository.deleteById(keywordUserMapId);
+
+        boolean isKeywordUsedByOthers = articleKeywordUserMapRepository.existsByArticleKeywordId(
+            articleKeywordUserMap.getArticleKeyword().getId());
+        if (!isKeywordUsedByOthers) {
+            articleKeywordRepository.deleteById(articleKeywordUserMap.getArticleKeyword().getId());
+        }
+    }
+
+    public ArticleKeywordsResponse getMyKeywords(Integer userId) {
+        List<ArticleKeywordUserMap> articleKeywordUserMaps = articleKeywordUserMapRepository.findAllByUserId(userId);
+
+        if (articleKeywordUserMaps.isEmpty()) {
+            return new ArticleKeywordsResponse(0, List.of());
+        }
+
+        return ArticleKeywordsResponse.from(articleKeywordUserMaps);
+    }
+
+}
