+++ src/test/java/in/koreatech/koin/fixture/ArticleFixture.java
@@ -4,13 +4,13 @@ import java.time.LocalDateTime;
 
 import org.springframework.stereotype.Component;
 
-import in.koreatech.koin.domain.community.model.Article;
-import in.koreatech.koin.domain.community.model.ArticleKeyword;
-import in.koreatech.koin.domain.community.model.ArticleKeywordUserMap;
-import in.koreatech.koin.domain.community.model.Board;
-import in.koreatech.koin.domain.community.repository.ArticleKeywordRepository;
-import in.koreatech.koin.domain.community.repository.ArticleKeywordUserMapRepository;
-import in.koreatech.koin.domain.community.repository.ArticleRepository;
+import in.koreatech.koin.domain.community.articles.model.Article;
+import in.koreatech.koin.domain.community.keyword.model.ArticleKeyword;
+import in.koreatech.koin.domain.community.keyword.model.ArticleKeywordUserMap;
+import in.koreatech.koin.domain.community.articles.model.Board;
+import in.koreatech.koin.domain.community.keyword.repository.ArticleKeywordRepository;
+import in.koreatech.koin.domain.community.keyword.repository.ArticleKeywordUserMapRepository;
+import in.koreatech.koin.domain.community.articles.repository.ArticleRepository;
 import in.koreatech.koin.domain.user.model.User;
 
 @Component
@@ -18,16 +18,12 @@ import in.koreatech.koin.domain.user.model.User;
 public class ArticleFixture {
 
     private final ArticleRepository articleRepository;
-    private final ArticleKeywordRepository articleKeywordRepository;
-    private final ArticleKeywordUserMapRepository articleKeywordUserMapRepository;
 
     public ArticleFixture(
         ArticleRepository articleRepository,
         ArticleKeywordRepository articleKeywordRepository,
         ArticleKeywordUserMapRepository articleKeywordUserMapRepository) {
         this.articleRepository = articleRepository;
-        this.articleKeywordRepository = articleKeywordRepository;
-        this.articleKeywordUserMapRepository = articleKeywordUserMapRepository;
     }
 
     public Article 자유글_1(User user, Board board) {
@@ -68,20 +64,6 @@ public class ArticleFixture {
                 .noticeArticleId(null)
                 .build()
         );
-    }
-
-    public ArticleKeywordUserMap 키워드1(String keyword, User user) {
-        ArticleKeyword articleKeyword = articleKeywordRepository.save(ArticleKeyword.builder()
-                .keyword(keyword)
-                .lastUsedAt(LocalDateTime.now())
-                .build());
-
-        ArticleKeywordUserMap articleKeywordUserMap = ArticleKeywordUserMap.builder()
-            .articleKeyword(articleKeyword)
-            .user(user)
-            .build();
-
-        return articleKeywordUserMapRepository.save(articleKeywordUserMap);
     }
 
     public ArticleFixtureBuilder builder() {
